Rename kcpforedgesyncer command to kcp_edge-syncer_gen and promote it to root command without having subcommands (#360)
