+++ pkg/cliplugins/kcp-edge/syncer-gen/edge-syncer.yaml
